Initial commit
This is the initial commit for my dwm-status tool. It will
replace the sh(1) script I'm using to update the dwm(1) status
bar at the moment.

At the moment it loops forever and prints the curent date and
time every three seconds to stdout.

I will go on implementing the basic features I need. Then I'll
replace the printing to stdout with the corresponding xsetroot(1)
call needed to update the dwm status bar.

Basic features means:

 * battery state
 * load averages