Standardize dwm-status.c
dwm-status is written in C89 but uses library functions only available in C99,
eg. snprintf(3).  This is the first in a row of commits to clean up things a
little.

Feature check macro

        #define _POSIX_C_SOURCE 199506

has been applied to dwm-status.c and the new tools.c to make sure that no
non-standard functions are used. I will go on and apply this to as much
functions as possible. I am not sure if this is feasible with all files but I'll
try and see how far I'll get.